text working, Barrier set up(Need Texture), bullets del when die

+++ Graphics Ass1/src/Sprite.h
@@ -32,7 +32,7 @@ public:
 	bool checkcollision(const float otherX, const float otherY,  float otherWidth,  float otherHeight);
 	void createBullet(vector<Sprite> &list, float direction);
 	
-	float health;
+	int health;
 	
 	float sizeH;
